+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/layout/activity_main.xml" value="0.2777777777777778" />
         <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/layout/activity_tutor_screen.xml" value="0.4" />
         <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/layout/activity_tutor_screen_actifivy.xml" value="0.209375" />
+        <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/layout/my_student_rv_item.xml" value="0.39791666666666664" />
         <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.242" />
         <entry key="..\:/MAP2/S2/6134/Project/TutorFinder/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.2165" />
       </map>
