+++ app/src/main/java/jwang/example/tutorfinder/tutor/EditProfileActivity.kt
@@ -1,7 +1,8 @@
-package jwang.example.tutorfinder
+package jwang.example.tutorfinder.tutor
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import jwang.example.tutorfinder.R
 
 class EditProfileActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
