+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfileActivity">
+    tools:context=".tutor.EditProfileActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
