+++ app/src/main/res/layout/activity_student_request.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudentRequestActivity">
+    tools:context=".tutor.StudentRequestActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
